requirements.txt
influxdb==5.0.0
python-dotenv==0.8.2
requests==2.18.4
sense-hat==2.2.0
urllib3==1.22
